requirements.txt
flask
torch
transformers
sentence-transformers
faiss-cpu
scikit-learn
pandas
datasets
tqdm
python-dotenv
